+++ lib/enums/dialog_status.dart
@@ -1,0 +1,11 @@
+/// Represents the current status of the dialog.
+enum DialogStatus {
+  /// Dialog is currently displayed
+  opened,
+
+  /// Dialog has been closed
+  closed,
+
+  /// Dialog has completed its progress
+  completed
+}