+++ lib/sn_progress_dialog.dart
@@ -4,3 +4,6 @@ library sn_progress_dialog;
 export 'progress_dialog.dart';
 export 'options/completed.dart';
 export 'options/cancel.dart';
+export 'enums/progress_types.dart';
+export 'enums/dialog_status.dart';
+export 'enums/value_position.dart';